backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-SocketIO
flask-cors
google-auth
google-auth-oauthlib
google-auth-httplib2
google-generativeai
python-dotenv
eventlet
